Moves LatLonCircle/Poly into their own files
